Time: 1234 ms (5.04%), Space: 403.8 MB (5.48%) - LeetHub

+++ 3800-longest-common-prefix-between-adjacent-strings-after-removals/3800-longest-common-prefix-between-adjacent-strings-after-removals.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     vector<int> longestCommonPrefix(vector<string>& words) {
-        if(words.size()==1)
+        if((int)words.size()==1)
             return {0}; 
 
         vector<int>v(words.size(),0);
